+++ src/command/commands/title.ts
@@ -1,12 +1,16 @@
 import { Platform } from '../../enums';
-import { ENV } from '../../config/config.service';
+import { CONFIG } from '../../config/config.service';
 
 const command: Command = {
   name: 'title',
   aliases: ['status'],
   platforms: [Platform.Twitch, Platform.Discord],
   run: async (ctx) => {
-    const url = `https://decapi.me/twitch/status/${ENV.TWITCH_CHANNEL}`;
+    if (!CONFIG.decapi.enabled) {
+      console.log('DecAPI is not enabled for !title command.');
+      return false;
+    }
+    const url = `https://decapi.me/twitch/status/${CONFIG.twitch.channel}`;
     const res = await (await fetch(url)).text();
     ctx.botSpeak(res);
     return true;