+++ src/command/commands/uptime.ts
@@ -1,17 +1,21 @@
 import { Platform } from '../../enums';
-import { ENV } from '../../config/config.service';
+import { CONFIG } from '../../config/config.service';
 
 const command: Command = {
   name: 'uptime',
   platforms: [Platform.Twitch, Platform.Discord],
   run: async (ctx) => {
-    const url = `https://decapi.me/twitch/uptime/${ENV.TWITCH_CHANNEL}`;
+    if (!CONFIG.decapi.enabled) {
+      console.log('DecAPI is not enabled for !uptime command.');
+      return false;
+    }
+    const url = `https://decapi.me/twitch/uptime/${CONFIG.twitch.channel}`;
 
     const res = await (await fetch(url)).text();
     if (res.includes('offline')) {
-      ctx.botSpeak(`${ENV.TWITCH_CHANNEL} is offline.`);
+      ctx.botSpeak(`${CONFIG.twitch.channel} is offline.`);
     } else {
-      ctx.botSpeak(`${ENV.TWITCH_CHANNEL} has been live for ${res}.`);
+      ctx.botSpeak(`${CONFIG.twitch.channel} has been live for ${res}.`);
     }
     return true;
   }