+++ src/command/owner-commands/bits.ts
@@ -1,7 +1,7 @@
 import { playAudioFile } from '../../utils/utils';
 import { FunctionQueue } from '../../utils/FunctionQueue';
 import { Platform } from '../../enums';
-import { YAML_CONFIG } from '../../config/config.service';
+import { CONFIG } from '../../config/config.service';
 
 const queue = new FunctionQueue();
 
@@ -20,16 +20,13 @@ const command: Command = {
         body = JSON.parse(body);
         const bits = parseInt(body.bits) || 0;
         let commands: string[], alert: string;
-        if (
-          YAML_CONFIG?.bitsConfig.matches &&
-          YAML_CONFIG.bitsConfig.matches[bits]
-        ) {
-          commands = YAML_CONFIG.bitsConfig.matches[bits].commands;
-          alert = YAML_CONFIG.bitsConfig.matches[bits].alert;
+        if (CONFIG?.bits.matches && CONFIG.bits.matches[bits]) {
+          commands = CONFIG.bits.matches[bits].commands;
+          alert = CONFIG.bits.matches[bits].alert;
         } else {
           const key = bits < 100 ? '99orLess' : '100orMore';
-          commands = YAML_CONFIG.bitsConfig[key].commands;
-          alert = YAML_CONFIG.bitsConfig[key].alert;
+          commands = CONFIG.bits[key].commands;
+          alert = CONFIG.bits[key].alert;
         }
         if (commands) {
           for (const cmd of commands) {