+++ src/command/owner-commands/raids.ts
@@ -1,10 +1,10 @@
 import { FunctionQueue } from '../../utils/FunctionQueue';
 import { playAudioFile, removeSymbols } from '../../utils/utils';
 import { Platform } from '../../enums';
-import { YAML_CONFIG } from '../../config/config.service';
+import { CONFIG } from '../../config/config.service';
 
 const queue = new FunctionQueue();
-const raidersConfig = YAML_CONFIG.raidConfig.matches || {};
+const raidersConfig = CONFIG.raids.matches || {};
 
 const command: Command = {
   name: 'raids',
@@ -27,16 +27,17 @@ const command: Command = {
             alert = raidersConfig[username].alert;
           }
         } else {
-          for (const cmd of YAML_CONFIG?.raidConfig.defaultCommands) {
+          for (const cmd of CONFIG?.raids.defaultCommands) {
             void services.twitchService.ownerRunCommand(cmd);
           }
-          alert = YAML_CONFIG?.raidConfig?.alert;
         }
-
+        if (!alert) {
+          alert = CONFIG.raids?.alert;
+        }
         if (alert) {
           await playAudioFile(alert);
         }
-        let announcement = YAML_CONFIG.raidConfig?.announcement;
+        let announcement = CONFIG.raids?.announcement;
         if (announcement) {
           announcement = announcement.replace(
             /{raider}/g,